+++ app/src/main/java/com/gd/form/activity/SearchTaskActivity.java
@@ -222,7 +222,7 @@ public class SearchTaskActivity extends BaseActivity {
     //获取作业区
     private void getArea() {
         JsonObject params = new JsonObject();
-        params.addProperty("employid", Integer.valueOf(userId));
+        params.addProperty("employid", Long.valueOf(userId));
         Net.create(Api.class).getSearchArea(token, params)
                 .enqueue(new NetCallback<List<SearchArea>>(this, true) {
                     @Override
